# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"success",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BusStatus": {
				"enum": [
					"IDLE",
					"MOVING",
					"STOPPED",
					"DELAYED",
					"OFFLINE"
				],
				"type": "string"
			},
...
			"CreateBusDTO": {
				"properties": {
					"collegeId": {
						"type": "string"
					},
					"busNumber": {
						"type": "string"
					},
					"vehicleNumber": {
						"type": "string"
					},
					"capacity": {
						"type": "number",
						"format": "double"
					},
					"driverId": {
						"type": "string"
					}
				},
				"required": [
					"collegeId",
					"busNumber",
					"vehicleNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateBusDTO": {
				"properties": {
					"busNumber": {
						"type": "string"
					},
					"vehicleNumber": {
						"type": "string"
					},
					"capacity": {
						"type": "number",
						"format": "double"
					},
					"driverId": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/BusStatus"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"TripType": {
				"enum": [
					"MORNING",
					"EVENING",
					"BOTH"
				],
				"type": "string"
			},
			"AssignBusRouteDTO": {
				"properties": {
					"busId": {
						"type": "string"
					},
					"routeId": {
						"type": "string"
					},
					"collegeId": {
						"type": "string"
					},
					"tripType": {
						"$ref": "#/components/schemas/TripType"
					}
				},
				"required": [
					"busId",
					"routeId",
					"collegeId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "access_token",
				"in": "cookie"
			}
		}
	},
	"info": {
		"title": "busapp-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/v1/buses": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "List all buses (filtered by college tenant).\n\n### Intent & Business Purpose\nFetches the list of registered buses assigned to a college campus. Used by fleet managers, admins, and tracking dashboards.\n\n### Target Audience & Roles\n- **Allowed Roles:** `ADMIN`, `COLLEGE_ADMIN`, `DRIVER`, `PARENT`, `STUDENT`.\n- **Access Control:** Requires valid JWT. College Admins and users are scoped strictly to their assigned `collegeId`. Super Admins can pass optional `collegeId` query parameter.\n\n### Query Parameters\n- `collegeId` *(optional string)*: Target college UUID. Overridden for tenant users.\n\n### Edge Cases & QA Testing Focus\n- Verify multi-tenant isolation: College A users cannot view buses from College B.\n- Returns empty array if no buses are registered for the tenant.",
				"tags": [
					"Buses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Optional college ID filter for Super Admins.",
						"in": "query",
						"name": "collegeId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Register a new fleet bus.\n\n### Intent & Business Purpose\nAdds a new bus vehicle to the college fleet inventory with registration number, capacity, and model info.\n\n### Target Audience & Roles\n- **Allowed Roles:** `ADMIN`, `COLLEGE_ADMIN`.\n- **Access Control:** Tenant College Admin or Super Admin.\n\n### Request Body\n- `busNumber` *(required string)*: Unique vehicle registration or plate number (e.g. `Ba 1 Pa 1234`).\n- `capacity` *(required number)*: Passenger seating capacity.\n- `model` *(optional string)*: Vehicle make/model.\n- `collegeId` *(required string)*: Associated college ID.\n\n### Side Effects\n- Inserts a new record into the `buses` table.\n\n### Edge Cases & QA Testing Focus\n- Enforces unique `busNumber` per college tenant. Duplicate registration number returns HTTP 400.",
				"tags": [
					"Buses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Bus registration details.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBusDTO",
								"description": "Bus registration details."
							}
						}
					}
				}
			}
		},
		"/api/v1/buses/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get bus details by ID.\n\n### Intent & Business Purpose\nRetrieves complete profile, current status, assigned driver, and assigned route details for a specific bus.\n\n### Target Audience & Roles\n- **Allowed Roles:** Authenticated users belonging to the tenant college.\n\n### Path Parameters\n- `id` *(required string)*: Unique UUID of the bus.\n\n### Edge Cases & QA Testing Focus\n- Accessing a bus ID belonging to a different college tenant returns HTTP 404 Not Found.",
				"tags": [
					"Buses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Unique bus UUID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update bus details.\n\n### Intent & Business Purpose\nUpdates vehicle information, status (`ACTIVE`, `MAINTENANCE`, `INACTIVE`), or capacity.\n\n### Target Audience & Roles\n- **Allowed Roles:** `ADMIN`, `COLLEGE_ADMIN`.\n\n### Path Parameters\n- `id` *(required string)*: Bus UUID.\n\n### Side Effects\n- Modifies existing `buses` table row.",
				"tags": [
					"Buses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Bus UUID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Fields to update.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBusDTO",
								"description": "Fields to update."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Soft-delete / deactivate a bus.\n\n### Intent & Business Purpose\nSoft-deactivates a bus record from active operational fleet views.\n\n### Target Audience & Roles\n- **Allowed Roles:** `ADMIN`, `COLLEGE_ADMIN`.\n\n### Path Parameters\n- `id` *(required string)*: Bus UUID.\n\n### Side Effects\n- Sets bus status to `INACTIVE` / soft deletes record.",
				"tags": [
					"Buses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Bus UUID.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/buses/assign-route": {
			"post": {
				"operationId": "AssignRoute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Assign bus to a specific transit route.\n\n### Intent & Business Purpose\nMaps a bus vehicle to a designated pickup/dropoff transit route (`routeId`).\n\n### Target Audience & Roles\n- **Allowed Roles:** `ADMIN`, `COLLEGE_ADMIN`.\n\n### Request Body\n- `busId` *(required string)*: Bus UUID.\n- `routeId` *(required string)*: Route UUID.\n- `collegeId` *(optional string)*: College UUID.\n\n### Side Effects\n- Updates `routeId` foreign key on the `buses` record.\n\n### Edge Cases & QA Testing Focus\n- Verify route belongs to the same college tenant.",
				"tags": [
					"Buses"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Assignment payload containing `busId` and `routeId`.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AssignBusRouteDTO",
								"description": "Assignment payload containing `busId` and `routeId`."
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:5000",
			"description": "Development server"
		}
	]
}